requirements.txt
python-telegram-bot
pymongo
pathvalidate